+++ app/src/main/java/ip/team13/petowner/core/helpers/AppConstants.kt
@@ -1,4 +1,4 @@
-package ip.team13.petowner.helpers
+package ip.team13.petowner.core.helpers
 
 object AppConstants {
     object SharedPrefKeys {
