+++ app/src/main/java/ip/team13/petowner/ui/activities/list/ItemActivityViewModel.kt
@@ -1,4 +1,4 @@
-package ip.team13.petowner.ui.activities.viewModels
+package ip.team13.petowner.ui.activities.list
 
 import ip.team13.petowner.data.domain.ActivityItem
