+++ app/src/main/java/ip/team13/petowner/ui/activities/list/ItemPetListViewModel.kt
@@ -1,0 +1,3 @@
+package ip.team13.petowner.ui.activities.list
+
+class ItemPetListViewModel()
