--- app/src/main/java/ip/team13/petowner/ui/activities/viewModels/ItemPetListViewModel.kt
@@ -1,3 +1,0 @@
-package ip.team13.petowner.ui.activities.viewModels
-
-class ItemPetListViewModel()
