+++ app/src/main/java/ip/team13/petowner/ui/register/RegisterViewModel.kt
@@ -3,7 +3,7 @@ package ip.team13.petowner.ui.register
 import androidx.databinding.ObservableField
 import androidx.lifecycle.ViewModel
 import ip.team13.petowner.data.repository.AuthRepository
-import ip.team13.petowner.helpers.isValidEmail
+import ip.team13.petowner.core.helpers.isValidEmail
 
 class RegisterViewModel : ViewModel() {
