+++ app/src/main/res/layout/item_activity_add.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="viewModel"
-            type="ip.team13.petowner.ui.activities.viewModels.ItemActivityAddViewModel" />
+            type="ip.team13.petowner.ui.activities.list.ItemActivityAddViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
