+++ app/src/main/res/layout/item_list_pets.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="viewModel"
-            type="ip.team13.petowner.ui.activities.viewModels.ItemPetListViewModel" />
+            type="ip.team13.petowner.ui.activities.list.ItemPetListViewModel" />
 
         <variable
             name="adapter"
